seperated menu into three parts and ran the app seperately
